News in restart and settings for WSL and fast restart
Added some more features, such as FASTER celery restart script with killing proc of celery and restart in threading manner, it allows me to kill\start celery in a 30 seconds, not few minutes as before.
Two examples, where _DEV is optimized for Windows Subsystem for Linux  and some settings with it.